Bigincrements keys changed for increments on migrations

+++ database/migrations/2019_11_17_235202_create_invoices_table.php
@@ -9,7 +9,7 @@ class CreateInvoicesTable extends Migration
     public function up()
     {
         Schema::create('invoices', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->timestamps();
             $table->dateTime('issued_at');
             $table->dateTime('overdued_at');
@@ -18,8 +18,8 @@ class CreateInvoicesTable extends Migration
             $table->double('total')->unsigned()->default(0);
             $table->string('status');
             $table->string('description')->nullable();
-            $table->unsignedBigInteger('client_id');
-            $table->unsignedBigInteger('provider_id')->nullable();
+            $table->unsignedInteger('client_id');
+            $table->unsignedInteger('provider_id')->nullable();
             $table->foreign('client_id')->references('id')->on('clients')->onDelete('cascade');
             $table->softDeletes();
         });
