+++ database/migrations/2019_11_18_032552_create_invoice_product_table.php
@@ -9,13 +9,13 @@ class CreateInvoiceProductTable extends Migration
     public function up()
     {
         Schema::create('invoice_product', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->timestamps();
             $table->unsignedInteger('quantity');
             $table->float('unit_price')->unsigned();
             $table->float('total_price')->unsigned();
-            $table->unsignedBigInteger('invoice_id');
-            $table->unsignedBigInteger('product_id');
+            $table->unsignedInteger('invoice_id');
+            $table->unsignedInteger('product_id');
             $table->foreign('invoice_id')->references('id')->on('invoices')->onDelete('cascade');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
         });
